https://github.com/stillgg/camp/issues/77
