+++ commercialNetwork/src/scss/career.scss
@@ -571,26 +571,53 @@
     }
   }
 
-  .error-message {
-    min-height: 26px;
-    height: 100%;
-    font-size: 14px;
-    font-style: italic;
-    color: $red;
-  }
-
   .file-info {
     min-height: 80px;
     height: 100%;
     display: flex;
     flex-direction: column;
+    transform: translateY(20px);
+    opacity: 0;
 
     .filename {
       display: block;
     }
 
-    .error-message {
+    &.active {
+      transition:
+        transform 0.3s ease,
+        opacity 0.3s ease;
+
+      transform: translateY(0px);
+      opacity: 1;
+
+      margin-top: 5px;
+      min-height: 80px;
+      height: 100%;
+      display: flex;
+      flex-direction: column;
+
+      .filename {
+        display: inline-flex;
+        align-items: center;
+        background-color: #d9d9d9;
+
+        &::before {
+          content: url("../resources/icons/paper.svg");
+          display: inline-block;
+          padding-right: 10px;
+          padding-top: 6px;
+        }
+      }
     }
+  }
+
+  .error-message {
+    min-height: 26px;
+    // height: 100%;
+    font-size: 14px;
+    font-style: italic;
+    color: $red;
   }
 
   &.active {
