+++ commercialNetwork/src/views/index.html
@@ -779,12 +779,7 @@
                             <input id="years" type="number" class="input__line" placeholder="Год" />
                           </div>
                         </div>
-                        <div
-                          data-error-message-days
-                          data-error-message-months
-                          data-error-message-years
-                          class="error-message"
-                        ></div>
+                        <div class="error-message"></div>
                       </div>
                       <div class="column__input">
                         <label class="input__title" for="tel">Контактный телефон</label>
